feat(shared-rasters): migrate compute_dem_derivatives to library mode (opt-in)
Moves the open-source TWI/FDR/FAC/slope/aspect pipeline (hybrid richdem +
WhiteboxTools) out of the standalone CLI script into a builder module under
gfv2_params.shared_rasters/ that the unified orchestrator can invoke when
explicitly opted in.

Status preserved: this pipeline produces a parallel artifact
(Twi_hydrodem_*.tif) and is NOT the canonical TWI source consumed by the
downstream PRMS parameter pipeline (carea_max/smidx_coef calibration
thresholds were tuned against the ArcPy-derived Twi_merged_*.tif from
merge_rpu_by_vpu_twi). The library docstring + a top-level "Status" note
in the module make this explicit; the orchestrator registers the builder
in BUILDERS + STEP_ORDER so it CAN run, but it is intentionally absent
from the default steps: list in configs/shared_rasters.yml. Users add it
to their own shared_rasters.yml to opt in.

All gotchas preserved verbatim from the original script:
- richdem float64 throughout the fill (epsilon-fill ULP gradient survival)
- WBT D8 routing instead of richdem's (int32 cell-index overflow at 2.147B
  cells on giant VPUs 03 and 10)
- richdem fill instead of WBT's (WBT --fix_flats doesn't terminate on
  continent-scale flats like VPU 18)
- Slope/aspect computed on the FIXED (pre-fill) DEM, not the filled one
  (avoids epsilon-fill imprint on flat-bottomed stream cells)
- LZW *without* predictor=2 on the fixed DEM (WhiteboxTools silently
  corrupts predictor=2 input; predictor=2 is safe only for the final TWI
  output consumed by GDAL-based tools)
- 60deg slope cap before tan() (eliminates negative TWI from spurious
  Hydrodem outlier cells; documented Adirondacks 5530 m artefact)

Legacy CLI shell (scripts/compute_dem_derivatives.py) preserves the
argparse interface and delegates to the library builder. Existing
slurm_batch/compute_dem_derivatives.batch keeps working unchanged.

PLANNED_STEPS grows to 9 entries; STEP_ORDER now has 4 (3 production +
1 optional). Smoke-verified that:
- Constants (DEM_NODATA, SLOPE_CAP_DEG) imported correctly
- Builder is invokable from the orchestrator when explicitly added to a
  step block; fails clearly with "Hydrodem source not found" when run
  against a VPU that has no input data

Step 2c of 6 on the consolidation branch.
